Renaming createError module to errorFactory
