+++ src/lib/server/errorFactory.ts
