+++ src/lib/server/mapApi.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from 'vitest';
 
 import type { ErrorType } from '$lib/types';
+import { createError } from '$lib/server/errorFactory';
 
-import { createError } from './createError';
 import { mapToApiError } from './mapApi';
 
 describe('mapToApiError', () => {
