+++ src/lib/server/services/noteService.ts
@@ -13,7 +13,7 @@ import {
 import { getUserByNoteId, getUser } from '$lib/server/db/userDb';
 import type { NoteEditor, NoteEditorInput, ServerError } from '$lib/types';
 import { generateId } from '$lib/identityGenerator';
-import { createError } from '$lib/server/createError';
+import { createError } from '$lib/server/errorFactory';
 import { sendEmail } from './emailService';
 
 export const updateNoteEditor = (input: NoteEditorInput): TE.TaskEither<ServerError, void> => {
